Add files via upload
The project consists of implementation of design patterns (Observer, Prototype and strategy) using java and unit testing is done using junit framework in eclipse sts.
It is deals with a small problem statement where a company's on boarding specialist handles the list of employees and their details and updates it from time to time. Also sick leaves of employee are stored and debit/credit to the salaries of the employee is done accordingly.